requirements_t1_gpu.txt
absl-py==0.5.0
astor==0.7.1
backcall==0.1.0
certifi==2018.8.24
colorama==0.4.0
cycler==0.10.0
Cython==0.29
decorator==4.3.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
opencv-python==3.4.3.18
pandas==0.23.4
parso==0.3.1
pickleshare==0.7.5
Pillow==6.2.2
prompt-toolkit==2.0.6
protobuf==3.6.1
pycocotools==2.0.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
tensorflow-gpu==1.15.2
tensorboard==1.11.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.3
wincertstore==0.2
